% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fle.R
\name{fle}
\alias{fle}
\title{Factor loading error}
\usage{
fle(A1, A2)
}
\arguments{
\item{A1}{A matrix of m rows and n columns.}

\item{A2}{A matrix of m rows and l columns where l can equal n.}
}
\value{
A numeric number
}
\description{
Computing the column space distance between two matrix
}
\examples{
fle(matrix(1:12, nrow=4), matrix(11:22, nrow=4));



}
