% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMSE.R
\name{qMSE}
\alias{qMSE}
\title{Quantile relative squared error}
\usage{
qMSE(x_true, x_est, q = 100)
}
\arguments{
\item{x_true}{True values, written in a vector of length n.}

\item{x_est}{Imputed or estimated values, written in a vector of length n.}

\item{q}{Number of partition intervals. If q equals n, then output is essentially relative mean squared error. Default is 100.}
}
\value{
A numeric number
}
\description{
Computing the q-quantile relative squared error as a generalised error measure on relative mean squared error
}
\examples{
qMSE(c(2, 3, 7, 1), c(-2, 0.5, 8, 2), 1);



}
