% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refold.R
\name{refold}
\alias{refold}
\title{Tensor refolding}
\usage{
refold(unfolding, k, dim_vec)
}
\arguments{
\item{unfolding}{A matrix.}

\item{k}{An integer specifying the mode of array to refold from.}

\item{dim_vec}{A vector specifying the expected dimension of output array.}
}
\value{
A multi-dimensional array
}
\description{
Performing to matrices tensorisation, which is the inverse process of unfolding
}
\examples{
refold(matrix(1:9,nrow=3), 1, c(3,1,3));


}
