% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttm.R
\name{ttm}
\alias{ttm}
\title{Mode k product with matrix}
\usage{
ttm(ten, A, k)
}
\arguments{
\item{ten}{A multi-dimensional array with the mode-k dimension m.}

\item{A}{A matrix with dimension n by m.}

\item{k}{An integer specifying the tensor mode to perform k-mode matrix product.}
}
\value{
A multi-dimensional array with the k mode dimension n
}
\description{
Performing k-mode matrix product of a tensor to a matrix
}
\examples{
ttm(array(1:24,c(3,4,2)), matrix(1:4,nrow =2), 3);



}
