% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{unfold}
\alias{unfold}
\title{Tensor unfolding}
\usage{
unfold(ten, k)
}
\arguments{
\item{ten}{A multi-dimensional array.}

\item{k}{An integer specifying the mode of array to unfold.}
}
\value{
A matrix
}
\description{
Performing to multi-dimensional arrays tensor unfolding, also known as matricization
}
\examples{
unfold(array(1:24, dim=c(3,4,2)), 2);



}
